Make it possible to compile package on Mac OS.
This requires conditional compilation of Linux-specific code.

This fixes https://github.com/brave-experiments/nitro-enclave-utils/issues/5.
