+++ system.go
@@ -1,11 +1,6 @@
 package enclaveutils
 
-import (
-	"net"
-	"syscall"
-
-	"github.com/milosgajdos/tenus"
-)
+import "syscall"
 
 const (
 	defaultFdCur = 65536
@@ -41,24 +36,4 @@ func setFdLimit(cur, max uint64) error {
 	elog.Printf("Modified file descriptor limit for cur=%d; max=%d.", rLimit.Cur, rLimit.Max)
 
 	return nil
-}
-
-// assignLoAddr assigns an IP address to the loopback interface, which is
-// necessary because Nitro enclaves don't do that out-of-the-box.  We need the
-// loopback interface because we run a simple TCP proxy that listens on
-// 127.0.0.1:1080 and converts AF_INET to AF_VSOCK.
-func assignLoAddr() error {
-	addrStr := "127.0.0.1/8"
-	l, err := tenus.NewLinkFrom("lo")
-	if err != nil {
-		return err
-	}
-	addr, network, err := net.ParseCIDR(addrStr)
-	if err != nil {
-		return err
-	}
-	if err = l.SetLinkIp(addr, network); err != nil {
-		return err
-	}
-	return l.SetLinkUp()
 }
