+++ system_darwin.go
@@ -1,0 +1,3 @@
+package enclaveutils
+
+func assignLoAddr() error { return nil }
