+++ system_linux.go
@@ -1,0 +1,27 @@
+package enclaveutils
+
+import (
+	"net"
+
+	"github.com/milosgajdos/tenus"
+)
+
+// assignLoAddr assigns an IP address to the loopback interface, which is
+// necessary because Nitro enclaves don't do that out-of-the-box.  We need the
+// loopback interface because we run a simple TCP proxy that listens on
+// 127.0.0.1:1080 and converts AF_INET to AF_VSOCK.
+func assignLoAddr() error {
+	addrStr := "127.0.0.1/8"
+	l, err := tenus.NewLinkFrom("lo")
+	if err != nil {
+		return err
+	}
+	addr, network, err := net.ParseCIDR(addrStr)
+	if err != nil {
+		return err
+	}
+	if err = l.SetLinkIp(addr, network); err != nil {
+		return err
+	}
+	return l.SetLinkUp()
+}
